Add files via upload
Traveling is an essential part of modern society. A trip to a neighboring city or aflight to a tropical island are some of the many travel options available that have becomeless expensive and stressful to plan in today's globalized world. This report defineshuman well-being as a multi-factor concept and describes its relationship withtourism. A literature review and the views of academic authors point to travel as anoptimal tool for stress management, emotional discharge, and inspiration. Most opinions haveshown that tourism is becoming an increasingly vital part of the ecosystem, whichinturn forms human well-being and satisfaction. Travel, in particular, diversifiesaccording to time and geography, while suppliers improve offers and take a customer-oriented approach. Most of the elements studied within the thesis are subjectiveconcepts, for example, inspiration and emotional well-being. Therefore, a qualitativeanalysis method and a sufficient gender-diverse sample were chosen to obtainprimaryinformation. Seventeen questions were asked to participants who, to varyingdegrees, described emotions, the paradigm of choice, expectations, and perspectives of aperson in relation to previous or future travels. The results proved the multi-layeredperception of human satisfaction, namely the definition of satisfaction betweenvarious people. In addition, the emotions and memories of traveling directly correlatewith expectations before the vacations. This study had some limitations, includingtheprospect of a larger sample and an increased number of questions. Nevertheless, thisisan optimal base for more detailed studies of human well-being and tourism, forexample, among ethnic or cultural minorities.